Move test to validateandinsertblock_test.go
